- Make an tensor.impl package that can be used from other tensor packages allowing bypass of defensive strategies   in the public interfaces. - Move private static inner classes TensorAddress.NumericTensorAddress and TensorAddress.StringTensorAddress   to tensor.impl package. - Use the StringTensorAddress.of from Reduce to avoid defensive array copy.
