+++ vespajlib/src/main/java/com/yahoo/tensor/impl/package-info.java
@@ -1,0 +1,6 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+@ExportPackage
+package com.yahoo.tensor.impl;
+
+import com.yahoo.osgi.annotation.ExportPackage;
